requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
PyPDF2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
